pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "single-cell-foundation-cta"
version = "0.1.0"
description = "Integrated CTA-only downstream finetuning repository for scBERT and scGPT."
requires-python = ">=3.10"
dependencies = [
  "anndata>=0.9",
  "einops>=0.6",
  "local-attention>=1.9",
  "matplotlib<3.9",
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "pytest>=8.0",
  "scikit-learn>=1.3",
  "scipy>=1.10",
  "torch>=2.1",
  "tqdm>=4.66",
]

[tool.setuptools.packages.find]
include = ["src*"]
